likes controller added, tested
