+++ app/views/posts/show.html.erb
@@ -4,11 +4,11 @@
     <p class="mt-3"><%= @post.author.full_name %><br><%= @post.created_at.to_time.utc.to_s %></p>
     <div class="like-unlike">
       <% if @post.likers.include? current_user %>
-        <%= link_to "<span class='glyphicon glyphicon-thumbs-up' aria-hidden='true'</span>".html_safe, like_post_user_path(id: current_user.id, post_id: @post.id), method: :post, class: 'btn btn-success', disabled: true %>
-        <%= link_to "<span class='glyphicon glyphicon-thumbs-down' aria-hidden='true'</span>".html_safe, unlike_post_user_path(id: current_user.id, post_id: @post.id), method: :delete, class: 'btn btn-danger' %>
+        <%= link_to "<span class='glyphicon glyphicon-thumbs-up' aria-hidden='true'</span>".html_safe, like_post_path(user_id: current_user.id, post_id: @post.id), method: :post, class: 'btn btn-success', disabled: true %>
+        <%= link_to "<span class='glyphicon glyphicon-thumbs-down' aria-hidden='true'</span>".html_safe, unlike_post_path(user_id: current_user.id, post_id: @post.id), method: :delete, class: 'btn btn-danger' %>
       <% else %>
-        <%= link_to "<span class='glyphicon glyphicon-thumbs-up' aria-hidden='true'</span>".html_safe, like_post_user_path(id: current_user.id, post_id: @post.id), method: :post, class: 'btn btn-success' %>
-        <%= link_to "<span class='glyphicon glyphicon-thumbs-down' aria-hidden='true'</span>".html_safe, unlike_post_user_path(id: current_user.id, post_id: @post.id), method: :delete, class: 'btn btn-danger', disabled: true %>
+        <%= link_to "<span class='glyphicon glyphicon-thumbs-up' aria-hidden='true'</span>".html_safe, like_post_path(user_id: current_user.id, post_id: @post.id), method: :post, class: 'btn btn-success' %>
+        <%= link_to "<span class='glyphicon glyphicon-thumbs-down' aria-hidden='true'</span>".html_safe, unlike_post_path(user_id: current_user.id, post_id: @post.id), method: :delete, class: 'btn btn-danger', disabled: true %>
       <% end %>
     </div>
     <div class="like-count justify-center mt-2 mb-2">
