+++ config/routes.rb
@@ -11,14 +11,16 @@ Rails.application.routes.draw do
     get 'posts', on: :member, to: 'users#user_posts'
     get 'friend_requests', on: :member, to: 'users#friend_requests'
     get 'friends', on: :member, to: 'users#friends_index'
-    post 'like_post', on: :member, to: 'users#like_post'
     post 'comment_post', on: :member, to: 'users#comment_post'
     post 'accept_friend_request', on: :member, to: 'users#accept_friend_request'
     post 'ask_friendship', on: :member, to: 'users#ask_friendship'
-    delete 'unlike_post', on: :member, to: 'users#unlike_post'
     delete 'reject_friend_request', on: :member, to: 'users#reject_friend_request'
     delete 'cancel_friendship', on: :member, to: 'users#cancel_friendship'
   end
+
+  # likes
+  post 'like_post', to: 'likes#like_post'
+  delete 'unlike_post', to: 'likes#unlike_post'
 
   resources :posts
 end
