+++ test/controllers/likes_controller_test.rb
@@ -1,0 +1,4 @@
+require 'test_helper'
+
+class LikesControllerTest < ActionDispatch::IntegrationTest
+end
