+++ test/integration/posts_likes_test.rb
@@ -29,41 +29,41 @@ class PostsLikesTest < ActionDispatch::IntegrationTest
 
   test 'user can like a post' do
     assert_difference('@post.likers.count') do
-      post(like_post_user_path(id: @user.id, post_id: @post.id))
+      post(like_post_path(user_id: @user.id, post_id: @post.id))
     end
   end
 
   test 'user can like a post only if he haven\'t liked the post yet' do
-    post(like_post_user_path(id: @user.id, post_id: @post.id))
+    post(like_post_path(user_id: @user.id, post_id: @post.id))
 
     assert_no_difference('@post.likers.count') do
-      post(like_post_user_path(id: @user.id, post_id: @post.id))
+      post(like_post_path(user_id: @user.id, post_id: @post.id))
     end
   end
 
   test 'like link is disabled if the user already liked the post' do
-    post(like_post_user_path(id: @user.id, post_id: @post.id))
+    post(like_post_path(user_id: @user.id, post_id: @post.id))
     get post_path(@post)
-    assert_select 'a[href=?][disabled="disabled"]', like_post_user_path(id: @user.id, post_id: @post.id)
+    assert_select 'a[href=?][disabled="disabled"]', like_post_path(user_id: @user.id, post_id: @post.id)
   end
 
   test 'user can unlike a post' do
-    post(like_post_user_path(id: @user.id, post_id: @post.id))
+    post(like_post_path(user_id: @user.id, post_id: @post.id))
 
     get post_path(@post)
-    assert_select 'a[href=?]', unlike_post_user_path(id: @user.id, post_id: @post.id)
+    assert_select 'a[href=?]', unlike_post_path(user_id: @user.id, post_id: @post.id)
 
     assert_difference('@post.likers.count', -1) do
-      delete(unlike_post_user_path(id: @user.id, post_id: @post.id))
+      delete(unlike_post_path(user_id: @user.id, post_id: @post.id))
     end
   end
 
   test 'user can\'t unlike a post if he haven\'t liked it yet' do
     get post_path(@post)
-    assert_select 'a[href=?][disabled="disabled"]', unlike_post_user_path(id: @user.id, post_id: @post.id)
+    assert_select 'a[href=?][disabled="disabled"]', unlike_post_path(user_id: @user.id, post_id: @post.id)
 
     assert_no_difference('@post.likers.count') do
-      delete(unlike_post_user_path(id: @user.id, post_id: @post.id))
+      delete(unlike_post_path(user_id: @user.id, post_id: @post.id))
     end
   end
 end
